// src/api/profile/types.ts
export interface ProfileUpdateData {
  first_name: string;
  second_name: string;
  display_name: string;
  login: string;
  email: string;
  phone: string;
}

export interface PasswordUpdateData {
  oldPassword: string;
  newPassword: string;
}

export interface AvatarUpdateResponse {
  id: number;
  first_name: string;
  second_name: string;
  display_name: string;
  login: string;
  email: string;
  phone: string;
  avatar: string;
}

export interface ProfileResponse extends AvatarUpdateResponse {}